+++ docs/CHANGELOG.rst
@@ -16,6 +16,9 @@ Added
 - Ubuntu 14.04/16.04 based Docker images for Resolwe
 - Add ``list_docker_images`` management command that lists all Docker
   images required by registered processes in either plain text or YAML
+- Data status is set to ``ERROR`` and error message is appended to
+  ``process_error`` if value of ``basic:storage:`` field is set to a
+  file with invalid JSON
 
 Changed
 -------